Compiler warnings: Conversion from size_t to int

+++ src/grib_accessor_class_gts_header.c
@@ -155,7 +155,7 @@ static int unpack_string(grib_accessor* a, char* val, size_t* len)
     grib_accessor_gts_header* self = (grib_accessor_gts_header*)a;
     grib_handle* h                 = grib_handle_of_accessor(a);
     int offset                     = 0;
-    int length                     = 0;
+    size_t length                  = 0;
 
     if (h->gts_header == NULL || h->gts_header_len < 8) {
         if (*len < 8)
